Renamed package, working on including Greengrass JAR
