+++ foundation/CDDBaselineJava/src/main/java/com/awslabs/aws/iot/greengrass/cdd/data/CddTopics.java
@@ -1,0 +1,32 @@
+package com.awslabs.aws.iot.greengrass.cdd.data;
+
+import com.awslabs.aws.iot.greengrass.cdd.providers.interfaces.EnvironmentProvider;
+
+import javax.inject.Inject;
+import java.util.Optional;
+
+public class CddTopics {
+    @Inject
+    private final EnvironmentProvider environmentProvider;
+    private Optional<String> cddBaselineTopic = Optional.empty();
+
+    public CddTopics(EnvironmentProvider environmentProvider) {
+        this.environmentProvider = environmentProvider;
+    }
+
+    private String buildCddBaselineTopic() {
+        if (!cddBaselineTopic.isPresent()) {
+            cddBaselineTopic = Optional.of(String.join("/",
+                    environmentProvider.getAwsIotThingName().orElseThrow(() -> new UnsupportedOperationException("Thing name missing from environment")),
+                    "cdd"));
+        }
+
+        return cddBaselineTopic.get();
+    }
+
+    public String getCddDriverTopic(Object object) {
+        return String.join("/",
+                buildCddBaselineTopic(),
+                environmentProvider.getDriverName(object.getClass()));
+    }
+}
