+++ foundation/CDDBaselineJava/src/main/java/com/awslabs/aws/iot/greengrass/cdd/events/GreengrassStartEvent.java
@@ -1,0 +1,7 @@
+package com.awslabs.aws.iot.greengrass.cdd.events;
+
+import org.immutables.value.Value;
+
+@Value.Immutable
+public abstract class GreengrassStartEvent {
+}
