+++ foundation/CDDBaselineJava/src/main/java/com/awslabs/aws/iot/greengrass/cdd/events/PublishBinaryEvent.java
@@ -1,0 +1,10 @@
+package com.awslabs.aws.iot.greengrass.cdd.events;
+
+import org.immutables.value.Value;
+
+@Value.Immutable
+public abstract class PublishBinaryEvent {
+    public abstract String getTopic();
+
+    public abstract byte[] getBytes();
+}
