+++ foundation/CDDBaselineJava/src/main/java/com/awslabs/aws/iot/greengrass/cdd/events/PublishMessageEvent.java
@@ -1,0 +1,10 @@
+package com.awslabs.aws.iot.greengrass.cdd.events;
+
+import org.immutables.value.Value;
+
+@Value.Immutable
+public abstract class PublishMessageEvent {
+    public abstract String getTopic();
+
+    public abstract String getMessage();
+}
