+++ foundation/CDDBaselineJava/src/main/java/com/awslabs/aws/iot/greengrass/cdd/events/PublishObjectEvent.java
@@ -1,0 +1,10 @@
+package com.awslabs.aws.iot.greengrass.cdd.events;
+
+import org.immutables.value.Value;
+
+@Value.Immutable
+public abstract class PublishObjectEvent {
+    public abstract String getTopic();
+
+    public abstract Object getObject();
+}
