+++ foundation/CDDBaselineJava/src/main/java/com/awslabs/aws/iot/greengrass/cdd/providers/interfaces/SdkErrorHandler.java
@@ -1,0 +1,7 @@
+package com.awslabs.aws.iot.greengrass.cdd.providers.interfaces;
+
+import com.amazonaws.SdkClientException;
+
+public interface SdkErrorHandler {
+    void handleSdkError(SdkClientException e);
+}
