--- foundation/CDDBaselineJava/src/main/java/com/timmattison/greengrass/cdd/communication/GreengrassCommunication.java
@@ -1,79 +1,0 @@
-package com.timmattison.greengrass.cdd.communication;
-
-import com.amazonaws.greengrass.javasdk.IotDataClient;
-import com.amazonaws.greengrass.javasdk.LambdaClient;
-import com.amazonaws.greengrass.javasdk.model.*;
-import com.google.gson.Gson;
-import com.timmattison.greengrass.cdd.providers.interfaces.EnvironmentProvider;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.inject.Inject;
-import java.nio.ByteBuffer;
-import java.util.Base64;
-import java.util.Map;
-import java.util.Optional;
-
-public class GreengrassCommunication implements Communication {
-    private final Logger log = LoggerFactory.getLogger(GreengrassCommunication.class);
-    public static final String EMPTY_CUSTOM_CONTEXT = "{}";
-    public static final String ENCODED_EMPTY_CUSTOM_CONTEXT = Base64.getEncoder().encodeToString(EMPTY_CUSTOM_CONTEXT.getBytes());
-    @Inject
-    private final IotDataClient iotDataClient;
-    @Inject
-    private final LambdaClient lambdaClient;
-    @Inject
-    private final EnvironmentProvider environmentProvider;
-
-    public GreengrassCommunication(EnvironmentProvider environmentProvider, LambdaClient lambdaClient, IotDataClient iotDataClient) {
-        this.environmentProvider = environmentProvider;
-        this.lambdaClient = lambdaClient;
-        this.iotDataClient = iotDataClient;
-    }
-
-    @Override
-    public void publish(String topic, Object object) throws GGIotDataException, GGLambdaException {
-        PublishRequest publishRequest = new PublishRequest().withTopic(topic).withPayload(ByteBuffer.wrap(new Gson().toJson(object).getBytes()));
-
-        iotDataClient.publish(publishRequest);
-    }
-
-    @Override
-    public void publish(String topic, byte[] bytes) throws GGIotDataException, GGLambdaException {
-        PublishRequest publishRequest = new PublishRequest().withTopic(topic).withPayload(ByteBuffer.wrap(bytes));
-
-        iotDataClient.publish(publishRequest);
-    }
-
-    public Optional<byte[]> invokeByName(String functionName, Optional<Map> customContext, byte[] binaryData) {
-        Optional<String> functionArn = environmentProvider.getLocalLambdaArn(functionName);
-
-        if (!functionArn.isPresent()) {
-            log.error("Attempted to invoke Lambda by name [" + functionName + "] but its ARN is not in the environment");
-            return Optional.empty();
-        }
-
-        String encodedClientContext = ENCODED_EMPTY_CUSTOM_CONTEXT;
-
-        if (customContext.isPresent()) {
-            encodedClientContext = Base64.getEncoder().encodeToString(new Gson().toJson(customContext.get()).getBytes());
-        }
-
-        final InvokeRequest invokeRequest = new InvokeRequest()
-                .withFunctionArn(functionArn.get())
-                .withInvocationType(InvocationType.RequestResponse)
-                .withClientContext(encodedClientContext)
-                .withPayload(ByteBuffer.wrap(binaryData));
-
-        try {
-            final InvokeResponse response;
-            response = lambdaClient.invoke(invokeRequest);
-            final byte[] bytes = response.getPayload().array();
-
-            return Optional.ofNullable(bytes);
-        } catch (GGLambdaException e) {
-            log.error("Lambda invoke failed [" + e.getMessage() + "]");
-            return Optional.empty();
-        }
-    }
-}
