--- foundation/CDDBaselineJava/src/main/java/com/timmattison/greengrass/cdd/events/GreengrassStartEvent.java
@@ -1,7 +1,0 @@
-package com.timmattison.greengrass.cdd.events;
-
-import org.immutables.value.Value;
-
-@Value.Immutable
-public abstract class GreengrassStartEvent {
-}
