--- foundation/CDDBaselineJava/src/main/java/com/timmattison/greengrass/cdd/events/PublishBinaryEvent.java
@@ -1,10 +1,0 @@
-package com.timmattison.greengrass.cdd.events;
-
-import org.immutables.value.Value;
-
-@Value.Immutable
-public abstract class PublishBinaryEvent {
-    public abstract String getTopic();
-
-    public abstract byte[] getBytes();
-}
