--- foundation/CDDBaselineJava/src/main/java/com/timmattison/greengrass/cdd/events/PublishMessageEvent.java
@@ -1,10 +1,0 @@
-package com.timmattison.greengrass.cdd.events;
-
-import org.immutables.value.Value;
-
-@Value.Immutable
-public abstract class PublishMessageEvent {
-    public abstract String getTopic();
-
-    public abstract String getMessage();
-}
