--- foundation/CDDBaselineJava/src/main/java/com/timmattison/greengrass/cdd/providers/interfaces/SdkErrorHandler.java
@@ -1,7 +1,0 @@
-package com.timmattison.greengrass.cdd.providers.interfaces;
-
-import com.amazonaws.SdkClientException;
-
-public interface SdkErrorHandler {
-    void handleSdkError(SdkClientException e);
-}
